requirements.txt
flask
flask-cors
pandas
joblib
numpy
scipy
google-generativeai
requests
beautifulsoup4
pdfminer.six
python-jobspy
scikit-learn
pdfplumber
python-docx
networkx
sentence-transformers
torch
gunicorn
